requirements.txt
# Financial Document AI
# Requirements for baseline experiments

# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.2.0

# Optional: OCR (for later phases)
# paddleocr>=2.6.0
# pytesseract>=0.3.10

# Optional: PDF processing (for later phases)
# pdf2image>=1.16.0
# PyMuPDF>=1.22.0
